+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Prop `openPerSection` to the `modal-trigger` block.
+- Prop `expiresDateCookie` to the `modal-trigger` block.
 
 ## [0.3.3] - 2020-04-06
 ### Fixed